# microservice/source/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Inventory API
  description: |-
    A demo inventory management system.
  version: 2.0.0
servers:
  - url: https://inventory-api.igraphql.co/api/v2
    description: Microservice
...
paths:
...
  /reviews/{skuId}:
    post:
      tags:
        - Reviews
      description: Creates a product review.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReview'
      responses:
        "201":
          description: Review successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode201"
        "400":
          description: Client sent an invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode400"
        "404":
          description: SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "409":
          description: Review already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode409"
        "415":
          description: Invalid Content-Type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode415"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    get:
      tags:
        - Reviews
      description: Retrieves a product's reviews.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
        - $ref: '#/components/parameters/Author'
        - $ref: '#/components/parameters/Field'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/ReviewID'
        - $ref: '#/components/parameters/Score'
        - $ref: '#/components/parameters/SummaryReview'
        - $ref: '#/components/parameters/UserID'
      responses:
        "200":
          description: Review(s) successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetReviews"
        "404":
          description: Review(s) or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    delete:
      tags:
        - Reviews
      description: Deletes a product's reviews
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
      responses:
        "200":
          description: Review(s) successfully deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "404":
          description: SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
  /reviews/{skuId}/{reviewId}:
    get:
      tags:
        - Reviews
      description: Retrieves a product review.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
        - name: reviewId
          in: path
          description: Review ID.
          required: true
          schema:
            type: integer
            example: 2
        - $ref: '#/components/parameters/Field'
      responses:
        "200":
          description: Review successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetReview"
        "404":
          description: Review or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    put:
      tags:
        - Reviews
      description: Updates a product review.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
        - name: reviewId
          in: path
          description: Review ID.
          required: true
          schema:
            type: integer
            example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReview'
      responses:
        "200":
          description: Review successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "400":
          description: Client sent an invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode400"
        "404":
          description: Review or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "415":
          description: Invalid Content-Type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode415"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    delete:
      tags:
        - Reviews
      description: Deletes a product review.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
        - name: reviewId
          in: path
          description: Review ID.
          required: true
          schema:
            type: integer
            example: 2
      responses:
        "200":
          description: Review successfully deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "404":
          description: Review or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
...
components:
  parameters:
...
    Author:
      name: author
      in: query
      description: Review author.
      schema:
        type: string
...
    Field:
      name: field
      in: query
      description: Limit the fields returned by the query. Can be specified multiple times.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
...
    Limit:
      name: limit
      in: query
      description: Limit the number of values returned by the query.
      schema:
        type: integer
...
    Offset:
      name: offset
      in: query
      description: Offset the number of values returned by the query.
      schema:
        type: integer
...
    ReviewID:
      name: review_id
      in: query
      description: Review ID.
      schema:
        type: integer
    Score:
      name: score
      in: query
      description: Review score.
      schema:
        type: number
...
    SummaryReview:
      name: summary
      in: query
      description: Review summary.
      schema:
        type: string
...
    UserID:
      name: user_id
      in: query
      description: Review User ID.
      schema:
        type: string
...
  schemas:
...
    CreateReview:
      type: object
      required:
        - author
        - score
        - summary
        - user_id
      properties:
        author:
          type: string
          example: Guest
        score:
          type: number
          example: 5
        summary:
          type: string
          example: Good, reasonably priced.
        user_id:
          type: string
          example: null
    GetReview:
      type: object
      properties:
        author:
          type: string
          example: Guest
        created:
          type: string
          example: 2023-07-04T12:00:00.00.000Z
        review_id:
          type: integer
          example: 2
        score:
          type: number
          example: 5
        sku_id:
          type: integer
          example: 96070
        summary:
          type: string
          example: Good, reasonably priced.
        updated:
          type: string
          example: 2023-07-04T12:00:00.00.000Z
        user_id:
          type: string
          example: null
    GetReviews:
      type: array
      items:
        type: object
        properties:
          author:
            type: string
            example: Guest
          created:
            type: string
            example: 2023-07-04T12:00:00.00.000Z
          review_id:
            type: integer
            example: 2
          score:
            type: number
            example: 5
          sku_id:
            type: integer
            example: 96070
          summary:
            type: string
            example: Good, reasonably priced.
          updated:
            type: string
            example: 2023-07-04T12:00:00.00.000Z
          user_id:
            type: string
            example: null
    UpdateReview:
      type: object
      properties:
        author:
          type: string
          example: Guest
        score:
          type: number
          example: 5
        summary:
          type: string
          example: Good, reasonably priced.
        user_id:
          type: string
          example: null
...
    HttpCode200:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: ok
    HttpCode201:
      type: object
      properties:
        status:
          type: integer
          example: 201
        message:
          type: string
          example: created
    HttpCode400:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: bad request
    HttpCode404:
      type: object
      properties:
        status:
          type: integer
          example: 404
        message:
          type: string
          example: not found
    HttpCode409:
      type: object
      properties:
        status:
          type: integer
          example: 409
        message:
          type: string
          example: conflict
    HttpCode415:
      type: object
      properties:
        status:
          type: integer
          example: 415
        message:
          type: string
          example: unsupported media type
    HttpCode500:
      type: object
      properties:
        status:
          type: integer
          example: 500
        message:
          type: string
          example: internal server error